+++ packages/@lou.codes/create-package/src/doneMessage.ts
@@ -5,13 +5,13 @@ import { userClean } from "./userClean.js";
 
 const git = foregroundGreen`git`;
 
-export const doneMessage = ({ name }: Answers) => `
+export const doneMessage = (answers: Answers) => `
 Done! ✨
 
-${foregroundGreen`cd`} ${underlined`./${name}`}
+${foregroundGreen`cd`} ${underlined`./${answers.name}`}
 ${git} init
 ${git} remote add origin ${underlined`git@github.com:${USER}/${userClean(
-	name,
+	answers.name,
 )}.git`}
 ${git} branch -M main
 ${foregroundGreen`npm`} i ${foregroundGray`# or pnpm or yarn`}